CV_v1/requirements.txt
scikit-learn
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.2.0
numpy==1.18.1
opencv-python==4.2.0.34
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
scipy==1.4.1
six==1.14.0
